Bugfix: reagent form2 render funcs need the same args as "parent"
Additionally, we need to move the let-binding calculations into the
main function, not outside it which would ignore any updated props.

This commit was actually made earlier, but because it wasn't pushed
and a more complicated commit was made in the same (now indented)
region, it was easier to drop that commit and just redo the change
manually.