process-layer: add pass-through behaviour if non-string is passed
This is a bit of a hack; it's because I want to re-use it later on, to
ensure a layer spec conforms, but in that case the bounding_box has
already been converted to a map.  Adding this hack simply means that
we don't try and convert a string to a map, if it's not a string any
more.