Calculate the max-height available for each layer group. Ref SM-34
This may not actually be the end of it; if a group only has a few
entries we can still get some pretty ugly entries.  The solution is a
bit more involved, I think; count the layers for each group, and
distribute things that way.  Both less-generic (each group needs to be
considered in isolation), and way fiddlier.  I think I'll leave it and
see how it plays out with real data.

The main case I'm worried about in practice is the bathymetry group:
that may only have a couple of entries, and is open by default.  But
the other two (how many image layers?  What if "other" is left open?)
can also throw a wrench into things.