project.clj tweak for figwheel: add cljs to :source-paths
That's an hour I'm not getting back: figwheel uses the top-level
project definition for clojure compilation, so a clj file for macro
definition that I had in the src/cljs directory wasn't being
compiled (although it *is* compiled by plain lein-cljsbuild).  Add
src/cljs to source-paths and everyone is happy.  (it's arguably more
correct to move the macro clj file to src/clj, but I'm happy with it
where it is)