Bugfix: handle errors when an element hasn't been (re)mounted yet
To be honest this is mainly of use during development, where a
figwheel-reload calls mount on everything again... and this time if
the helper layer is present it starts trying to get element dimensions
of things that may not have been remounted yet, and explodes.  Don't
explode, just skip that element for now.

(I guess it's not ideal; it won't render the help text for that
element, but it also won't then update when they are mounted, because
the props of the component won't have changed.  It does ease
development pains though)