Functional popup: add React key, and actual content
The main change is the addition of a react key based on the location.
There was a bug previously where any other click would close the
popup,  but not otherwise re-open it with the new subscription info;
the dynamic position props were updated but the state was still
"closed".  This forces creation of a new one each time.

Secondly, we dangerouslysetinnerhtml for the popup content.