A show-messages component, that uses the Blueprint toaster component
This was a tad trickier (I tried a slightly nicer approach first using
changing props to break up subscription and presentation, but it
didn't work so this will do); the toaster actually uses a static
create method and another to display the toast, rather than a regular
react component.  So basically we listen on the subscription, call
that method when we get a new value (the handler makes sure there's a
cache-busting token to avoid duplicate messages being hidden), and
return nil. Ugh.