Significantly rework segment-reassembly algorithm. Ref Closes-55
... But there's still a significant issue around tolerance and
precision.  The main culprit in testing is the
SeamapAus_NSW_marine_habitats_2013 table, which has 47k polygons and
nearly 400 less than a metre-square, causing a lot of numerical
instability.  Because we have to do point-equality, things get risky
as soon as we start simplifying too.

Probably needs simplification of the underlying dataset.

Nonetheless, the original issue of algorithm issues is resolved, so
yay?