Make sure we update map props on load. Closes SM-74
Recap: we now filter layers shown to only those in the viewport, and
filtering is done using the map bounds.. but those are only populated
when the user interacts with the map, so we need something to fire
when the map first loads.  This got tricky, because "load" never fires
-- it turns out, because the handler is only attached after the map is
already constructed (and loaded).  So, we use layeradd instead, which
fires when the base-layer is added, and obviously again later but it's
still low-overhead.

(note: the map itself offers the whenReady() method which fires its
callback arg exactly once, but we don't have access to that)