Refactor: use raw habitat data, rather than preprocessing in sub
Joe's original code assumed three-tuples of start/end/name, so there
was a simple bit of processing in the event handler to turn the
sequence of maps from the server into that format.  We need extra
information from it, and there's no real benefit from doing it anyway,
so refactor the code to use the native format now.