Bugfix: get mouse x/y from containerPoint, not layerPoint
Closes SM-60.  Man I'm glad that wasn't something harder like numeric
instability; just a stupid mistake (layerpoint changes as the layer is
moved around, zoomed etc, and can go negative)

It was confusing though because the main "test" was to click
zoom-to-extent, at which point the layerPoint and containerPoint
coords were both about the same.  Then a click typically causes an
autopan, changing layercoords, and pretty soon there's no "results",
repeatably.