Force habitat polygons in order of layer-argument. Closes SM-101
Previously we selected polygons using a "where layer in (...)"
statement, where the layer ordering determined the ordering in the
argument list.  This did not unfortunately guarantee that the result
polygons would have the same ordering though, which is what we rely on
to guarantee correct clipping / overlaps.

Instead, we generate a separate select statement for each layer, and
union-all them together, along with a priority column generated from
the order of layers, then order by that.