Tooltip over clipped-text, not full text. Closes SM-102
Problem here was the tooltip would appear to the right of the *full*
text, ie out in the middle of the map somewhere (and at a different
location depending on the length of the text itself).  This makes the
position consistent, and relative to the clipped-text component.

Note, css change needed to ensure the width is still calculated
correctly for the clipper component.