Typoes in saved view file. Ref SM-112
That was weird; there was an invisible unicode in the Vic WPB line
which was causing the bug (because the name didn't actually match),
but also for some reason my saved file wasn't even valid because there
was a union-all line in the wrong place.  Anyway, it /looks/ ok now,
and the database is up to date so things work again.  Closing the
ticket.