Add :put-state effect and hook up. Ref SM-115
A simple effect that sets an arbitrary string in the URL hash.  Hooked
up so when any state parameters of interest change the URL is updated.
I'm not sold on this approach as it's a bit fragile; it had to touch
the code in a few places, so if any was missed the URL could be
inaccurate, and any parameters later added to the hash will also
require extensive changes.

The other approaches I was considering were lighter-touch, but heavier
on side-effects: a component that only rendered null but did subscribe
to the state-value and set the URL that way, or possibly (didn't
investigate this far) a middleware/interceptor for reframe that I
think might have worked.

Will see how this works for now.