Make transactions cancelable. Closes SM-128
The gist here is we maintain a query id which is both stored in the db
and threaded through ajax handlers; in the response callbacks we check
if the associated query id is the same as the "current" (app-db) id,
and ignore that event if it is.

There's a tiny bit of thought around cancelling, because the most
common use-case is actually probably going to be where bathymetry has
promptly returned, but habitat is still going.  Ie, while it's easy to
put a cancel button on the "loading" screen, we mostly want it on the
"partial" view.  This is implemented with a bit of extra logic in the
query-cancel handler, to work in conjunction with the results-state
sub, to clear :loading results but leave everything else.