Help-layer can use element props OR map descriptors
Previously we just had an id, and all text to display etc was taken
from attributes on that element.  The context for this change is the
layer groups, which are very similar and it's going to get pretty
clumsy specifying helper text in each individual entry.

Possible future enhancement: allow map keys to be specified in
clojure-form, and converted to camelCase as produced by the
attribute-extraction process (and thus expected by processing)