Adjust height when embedded in wordpress. Closes SM-177
That completely hides where the rest of the day went, when I realised
that the app also became massively laggy when embedded.  This
was *eventually* tracked down to a catchall (selector: *) "make all
transitions like this", which promptly slowed leaflet to a crawl, for
the sin of using translate3d in its tiles.  Good lord -- I'm never
getting that day back, but the chrome dev tools are utterly amazing.