Redo popup response parsing using js/DOMParser
A bit unfortunate, but it seems pretty well supported (and we already
use at least one other "modern" API that escapes me right now).  The
motivation is the new popup for habitat-stats I'm introducing, which
includes a url with query parameters, and the "&" in eg ?p1=one&p2=...
made the old xml parser explode.