Expand extra-params sub; fade other layers when downloading.
Closes SM-197

Need to be careful here; if we remove the CQL_FILTER parameter, it
doesn't seem to count as a different prop for the purposes of
update (I /think/ that's what was happening... I'm doubting now that I
type it out though).  So we always specify an opacity, and only
override it when required.  For boundary layers we always specify a
CQL_FILTER, but effectively use a no-op most of the time.