Boundary layers can be used in conjunction with other layers. Closes SM-130
Needed a bit of thought, but a small change overall.  The gist is I'd
previously made bathymetry layers exclusive from other layers, because
the heatmap-style was so heavy it made no sense to display in
conjunction with others.  A contour layer is different though; it's
fairly sparse, so can be usefully displayed on top of habitats.

There's a bit of sleight-of-hand in the implementation: we actually
introduce a new category (note that fixtures are updated too)
"contours", and at load-time we map contours to bathymetry, but with
an additional attribute to identify them as a special type of bathy
layer, and base our logic on that accordingly.

There's actually a small related bugfix included; previously turning a
bathy layer on would remove other layers, but you could still add a
layer on top of a bathy layer... this has been fixed.