Remove overflow:hidden from body element
This was originally in place from the stand-alone page, but it has
since been included in the wordpress template... meaning pages aren't
scrollable!  Should be fine to add that property to the
application (#app) itself instead -- removing it altogether doesn't
seem to have any effect, and shouldn't, but I'm leaving it in as a
safety measure.  (yes, voodoo CSS programming, like all my CSS work)