Handle POINTs in subset data. Closes SM-246
There's actually a few parts to this.  One is that we need to handle
points, addressed here.  The other is that the *actual* issue turned
out to be an additional geometry column in the table (used for
zoomed-out map layers), containing points, and because we had an
implicit assumption there was only a single geometry field the
detection and handling was fairly naive (in this case, the geometry
column being picked was actually the one containing just points).  We
could have well-known-named-columns, but apart from being fragile it
doesn't even really make sense to have more than one geom column, so
we'll handle it as data instead (ie, the layer configuration will
change)