Bugfix: no effects if filtered layers list is empty. Closes SM-249
When a click is handled, we first filter out the bathymetry layers,
but somehow missed the case where there were /only/ bathymetry layers
active; check we have layers remaining before processing.  Note, this
still leaves the desirable case where a contour bathy layer is
displayed on top of a habitat layer, but we're interested in the
featureInfo for the habitat layer.