ISA-12: Add a loading screen while image is being prepared
This did get a bit messy: the adapt-react-class wrapper allows for the
use of clojure-style keywords for props and event handlers, eg
on-moveend gets translated to onmoveend.  Secondly, react-leaflet
lowercases events specified via react props.  Thirdly, the easyprint
plugin uses *both* camel-case *and* hyphens: easyPrint-start, etc.  So
we can't bind it via a react prop, and instead it is horribly hacked
by a :ref handler, and adding directly to the leaflet element.  But
that's not all, because that will be invoked every time the component
is updated, so the final step is a bit of (again dipping into native
js) hackery to make sure it is only added once (assuming no-one else
wants to listen to those events!)