ISA-19: uniform WMS version specifier
Well that summary, and the mostly-whitespace or irrelevant diff, hides
a nasty bug: rendering tiles using version 1.1.1 but getfeatureinfo
using 1.3.0, seems to have messed up precision horribly.  I don't
undrestand why and I might be hiding the real cause, but I haven't
gone to the extra effort of truly refining the test case.  It did seem
to be the differentiator though!  In the process I'm also removing
parameters that aren't required for that version, and minor
corrections.