ISA-20: Third-party tree-view state should be robust to renaming
The serialised state used the org/category *name* in the state
identifiers, meaning they would lose that state if the org or category
was renamed.  Tweak to add the id as well, so the open/collapsed state
is robust across renaming.

Note this requires a redeploy of the API backend as well.