WIP: clean up namespaces and unused parameters
This is mainly to keep clj-kondo (/lsp) a bit quieter, and while we do
a refresh.  Often I have used an "_" prefix as a form of
documentation, particularly when a vector (eg [top bottom left right])
is passed in where only a couple of elements are used, or :keys
map-destructuring is used an the :as alias is retained for
documentation even if never used.