ISA-70: WIP: Manually add a selection of layers
Next step is to turn this into a sub, with state controlled by
re-frame.  In particular, I think this is causing issues because
changing is causing a map.addLayer() call under the hood, which is
appearing on top.  May need to manually alter z-index, or simply
rendering the state on selection change should be fine.