ISA-85: handle startup corner-case involving empty map bounds
Unlike zoom level etc, the map's geographic bounds are populated by
leaflet events.  At startup, we haven't received any (technically, we
haven't processed any -- due to other corner cases we ignore map
events until application has loaded), so the bounds are empty.  This
would mean that when we are filtering the list of applicable layers
they would *all* be considered not intersecting, so if we have no
bounds object we consider that an intersection.