ISA-74: Increased tolerance for tile-load errors
We used to keep track of which layers had seen any tile-load error,
and display a warning (even if it was a single glitch over a long
time).

We have replaced this with a count of tile-load starts (ie, count of
tiles requested), and a separate count of tile errors (both per-layer
of course).  We then consider a layer as problematic if the percentage
of errors is greater than a certain tolerance, currently 40% (future
work: potentially make that configurable)