ISA-99: shadow-cljs config :init-fn -> :entries
Our (relatively recent) shadow-cljs setup used an :init-fn to indicate
the entry point, which I had assumed was about ensure the function
name was stable and not mangled. As it turns out, it actually adds
a *call* to that function at the end of the module, which in
conjunction with the traditional invocation in the web page was
calling init() twice (the side-effects of which were a bit
browser-dependent)

The two options for resolution were to remove the in-page call to
init() (ie, just load the module and let it proceed), or to remove the
invocation from the module. We decided on the latter, treating the
module more like a library, and giving us control over when it
launches.