ISA-417: relax metadata-url match regexp
We use a regexp to match metadata urls to determine whether they're
going to contain a format we recognise and can parse (and hence can
show the abstract in a modal), or if we just show the default
no-detail modal with a link for people to look for themselves.

Previously this matched UUIDs,  specifically version 4, but we have a
few records that didn't match (I assume, IDs were created externally
then re-hosted).  Since it's not that important to be strict we just
relax the check.