Build update: don't need if-develop/if-master different paths now
Context: the app structure and build process diverged a bit while v2
was in development, so a few steps in the build script needed to check
which branch was active.  Since merging develop into master, this is
no longer necessary.

(A consequence of this though is that a recent production deployment
failed because a build step had silently failed because of previous
assumptions)