ISA-425: reset renderer to avoid exceptions causing more exceptions
Issue here, as outlined in the comment, is that once an exception from
a view has been processed (in this case, a missing-parameter
exception), it is still passed on to the custom renderer that may have
been set.

Some built-ins explicitly check for exceptions (eg,
StaticHTMLTemplateRenderer), but we elect to just reset to json and
let the user sort it out.