ISA-429 / ISA-416: rework legend dispatch to accommodate all servers
I think we have all bases covered here.  The gist is that we assume a
wms server is either geoserver (supporting json legends), or ESRI in
which case we can differentiate based on the URL (contains MapServer).
ESRI feature server layers are differentiated based on layer type.

We remove the fallback error handling.