ISA-431: reproduce hierarchical catalogue in in-region drawer
Essentially we reuse the same component, but it necessitated a few
refactorings to enable this.  Namely, instead of a simple db path, we
add a "catalogue id" to the path, so now events and subs can reference
the catalogue they refer to, meaning nodes can have separate expansion
states between different catalogues, etc.