ISA-473: experimental worker to parse ESRI data into geojson
The issue seems to be that the response taking a long time (eg 20s) to
return is fine, but the parsing / conversion to geojson can also take
quite an appreciable amount of time, and since that happens inside the
xhr callback it all blocks up the task queue and locks up the browser.
We spawn the parsing and conversion into a separate thread (a Worker),
and the remaining callback is simple the callback invoked on the
worker's response message, freeing up resources.

Note that there's no thread management (such as a pool) other than
cleaning up each one after it has finished, and the processing is
still pretty heavy on resources.

To achieve all this: we vendor the esri-leaflet library and make that
small change to use a worker, then write a worker that is just the
parsing code mentioned (mostly from @terraformer/arcgis)

The bigger complications are integrating into seamap's build process,
using shadow-cljs.  Shadow has support for building workers, but there
are issues with hashed module names for eg.

We may yet introduce parceljs or similar to simplify the worker
compilation, as it seems more efficient than shadow.