ISA-473: properly ensure tiles only requested once metadata available
This was insidious: ESRI supports geojson out of the box (well, recent
servers, and where there's no "M" attribute), *but* the asynchrony
meant that tiles were requested at the same time as the initial
metadata query, so they didn't have the "geojson available"
information.

The basic fix was moving the call to _update (which loads tiles) from
onAdd() to the callback from metadata-received, but there are a few
other ways it can be invoked and we don't have a great flag for
"metadata received", so we make do.