ISA-637: WIP but promising bounding-box reprojection
Essentially: coordinates are stored geographically, which obviously
has issues around the poles, but leaflet otherwise handles it. The
issue really arises when we want to do reason about (do calculations
on) the extent represented by that bounding box: for something
covering the entire pole, the left and right edges meet (form a single
line) so simply taking the mins and maxes does nothing.

What we need is the geographic *coordinates* of the points
representing the lower-left and upper-right corners of the bounding
box in *viewport* space.

There may be a simpler way, but the code I adapted samples along each
edge, transforms each point and takes the min and max of x and y, then
unprojects the resulting corner coordinates from projected space.
This gives geographic coords that aren't necessarily the min or max in
terms of extents, but are the min and max of the viewport area of
interest.

It remains to be seen what new problems this introduces.