Adding new implementation of bounds-for-zoom based on openlayers
We add bounds-for-zoom2 for now.

(the old implementation was also based on openlayers I think, or
perhaps just the geoserver implementation of openlayers preview)

The previous approach was to linearly scale the map bounds to a known
pixel bounds.  This doesn't work on polar coordinates!  Instead we
take the resolution (or the inverse of scale; both are properties of
the zoom) and use that to construct the bounding box, ignoring the map
dimensions altogther.  This is more stable because it's in a smaller
neighbourhood.

Unfortunately: not quite ready for prime-time because the
"scale(zoom)" method in the default projections is not the same as the
one where you construct a CRS from an array of scales, which we do for
EPSG3031.