Improve: bounding box calculation for polar layers
We had cribbed some nice code to create a series of points around the
edge of a rectangle and transform each, but the weakeness was assuming
that for eg the minimum x value would come from the left edge.  This
doesn't apply to fully circumpolar layers (basically all min AND maxes
are going to come from a single edge I think!), so transform each
point as before, but now calculate the mins and maxes completely
separately and reconstruct the bounding box from these.