Bugfix: customise the initialise-layers boot to not request boundaries
This was causing errors in dev (masked because I'd commented it out
when I was first getting started and then forgotten about it) because
the tables were missing so the URLs would 500 and the site not load.

Note for future archeologists: there's quite a bit of reused
functionality in this (Antarctica) module, which may cause more
conflicts later on.  It would be nice to have plugin-driven more
modular structure, but that's a bit of a refactor / redesign away.